requirements.txt
streamlit
keras.models
PIL
numpy
streamlit_option_menu
proses
streamlit_webrtc
cv2
base64
gtts